pyproject.toml
[project]
name = "mindframe"
version = "0.2.1"
description = "A Python package for the `mindframe` project"
requires-python = ">=3.11"
authors = [
    {name = "Ben Whalley", email = "ben.whalley@plymouth.ac.uk"},
    {name = "Ali Bidaran", email = "ali@example.com"},
]

dependencies = [
  "black",
  "boto3>=1.36.25",
  "celery>=5.4.0",
  "colorlog>=6.9.0",
  "crispy-bootstrap5>=2024.10",
  "dateparser>=1.2.1",
  "dj-database-url>=2.3.0",
  "django>=5.1.6",
  "django-autoslug>=1.9.9",
  "django-celery-beat>=2.7.0",
  "django-compressor>=4.5.1",
  "django-crispy-forms>=2.3",
  "django-debug-toolbar>=5.0.1",
  "django-environ>=0.12.0",
  "django-extensions>=3.2.3",
  "django-hijack>=3.7.1",
  "django-lifecycle>=1.2.4",
  # "django-magiclink",
  "django-ninja>=1.3.0",
  "django-shortuuidfield>=0.1.3",
  "django-storages>=1.14.5",
  "django-treebeard>=4.7.1",
  "djmail>=2.0.0",
  "gradio>=5.17.0",
  "haystack-ai>=2.10.3",
  "instructor>=1.7.2",
  "ipython>=8.32.0",
  "langchain>=0.3.19",
  "langfuse>=2.59.3",
  "nltk>=3.9.1",
  "pandoc>=2.4",
  "pgvector>=0.3.6",
  "pgvector-haystack>=2.0.0",
  "poetry-plugin-export>=1.9.0",
  'psycopg2 ~=2.9; sys_platform == "linux"',
  'psycopg2-binary ~=2.9; sys_platform == "darwin"',
  "pydantic>=2.10.6",
  "python-box>=7.3.2",
  "python-decouple>=3.8",
  "python-magic>=0.4.27",
  "python-telegram-bot>=21.10",
  "recurrent>=0.4.1",
  "redis>=5.2.1",
  "requests>=2.32.3",
  "ruamel-yaml>=0.18.10",
  "rules>=3.5",
  "setuptools>=75.8.0",
  "shortuuid>=1.0.13",
  "tiktoken>=0.9.0",
  "uvicorn>=0.34.0",
  "watchdog[watchmedo]>=6.0.0",
  "watchfiles>=1.0.4",
  "werkzeug>=3.1.3",
  "whitenoise>=6.9.0",
]

[tool.uv.sources]
# django-magiclink = { git = "https://github.com/benwhalley/django-magiclink.git" }

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
